demos/requirements.txt
google-cloud-firestore==2.15.0
google-cloud-bigquery==3.20.1
